\name{timezero-package}
\alias{timezero-package}
\alias{timezero}
\docType{package}
\title{
\packageTitle{timezero}
}
\description{
\packageDescription{timezero}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{timezero}
\packageIndices{timezero}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{timezero}

Maintainer: \packageMaintainer{timezero}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
